Lab 6

+++ Pokemon/src/main/java/org/example/pokemon/service/PokemonService.java
@@ -1,7 +1,12 @@
 package org.example.pokemon.service;
 
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.ejb.EJBAccessException;
+import jakarta.ejb.LocalBean;
+import jakarta.ejb.Stateless;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import jakarta.security.enterprise.SecurityContext;
 import jakarta.transaction.Transactional;
 import jakarta.ws.rs.NotFoundException;
 import lombok.NoArgsConstructor;
@@ -10,28 +15,40 @@ import org.example.pokemon.dto.request.PokemonCreateRequest;
 import org.example.pokemon.dto.request.PokemonEditRequest;
 import org.example.pokemon.dto.response.PokemonResponse;
 import org.example.pokemon.entity.Pokemon;
+import org.example.pokemon.entity.PokemonSpecies;
+import org.example.pokemon.entity.User;
+import org.example.pokemon.entity.UserRole;
 import org.example.pokemon.repository.impl.PokemonH2Repository;
 import org.example.pokemon.repository.impl.SpeciesH2Repository;
+import org.example.pokemon.repository.impl.UserH2Repository;
+
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-@ApplicationScoped
+@LocalBean
+@Stateless
 @NoArgsConstructor
 public class PokemonService {
 
     private PokemonH2Repository pokemonH2Repository;
     private SpeciesH2Repository speciesH2Repository;
+    private UserH2Repository userH2Repository;
     private DtoFunctionFactory factory;
+    private SecurityContext securityContext;
 
     @Inject
-    public PokemonService(PokemonH2Repository pokemonH2Repository, SpeciesH2Repository speciesH2Repository, DtoFunctionFactory factory) {
+    public PokemonService(PokemonH2Repository pokemonH2Repository, SpeciesH2Repository speciesH2Repository, UserH2Repository userH2Repository ,DtoFunctionFactory factory,
+                          SecurityContext securityContext) {
         this.pokemonH2Repository = pokemonH2Repository;
         this.speciesH2Repository = speciesH2Repository;
+        this.userH2Repository = userH2Repository;
         this.factory = factory;
+        this.securityContext = securityContext;
     }
 
-    @Transactional
+
     public void create(Pokemon pokemon){
         if(pokemonH2Repository.find(pokemon.getId()).isEmpty())
             pokemonH2Repository.create(pokemon);
@@ -46,16 +63,22 @@ public class PokemonService {
         pokemonH2Repository.create(pokemonToCreate);
     }
 
-    @Transactional
+
     public void create(UUID speciesId, PokemonCreateRequest pokemonRequest) {
         var species = speciesH2Repository.find(speciesId)
                 .orElseThrow(() -> new IllegalArgumentException("Pokemon species with id " + speciesId + " not found"));
         var pokemonToCreate = factory.pokemonCreateRequestToPokemon().apply(pokemonRequest);
         pokemonToCreate.setSpecies(species);
+
+        User user = userH2Repository.findByUsername(securityContext.getCallerPrincipal().getName())
+                .orElseThrow(IllegalStateException::new);
+
+        pokemonToCreate.setOwner(user);
         pokemonH2Repository.create(pokemonToCreate);
     }
 
     public void update(PokemonEditRequest pokemonRequest) {
+        checkAdminRoleOrOwner(pokemonH2Repository.find(pokemonRequest.getId()));
         pokemonRequest.setSpeciesId(pokemonRequest.getSpeciesId());
         var species = speciesH2Repository.find(pokemonRequest.getSpeciesId())
                 .orElseThrow(() -> new IllegalArgumentException("Pokemon species with id " + pokemonRequest.getSpeciesId() + " not found"));
@@ -64,8 +87,8 @@ public class PokemonService {
         pokemonH2Repository.update(pokemonToUpdate);
     }
 
-    @Transactional
     public void update(UUID speciesId, UUID pokemonId, PokemonEditRequest pokemonRequest) {
+        checkAdminRoleOrOwner(pokemonH2Repository.find(pokemonId));
         pokemonRequest.setId(pokemonId);
         var species = speciesH2Repository.find(speciesId)
                 .orElseThrow(() -> new NotFoundException("Pokemon species with id " + pokemonRequest.getSpeciesId() + " not found"));
@@ -81,13 +104,23 @@ public class PokemonService {
     }
 
     public List<PokemonResponse> getPokemonsBySpeciesId(UUID speciesId) {
-        System.out.println(pokemonH2Repository.findAllBySpeciesId(speciesId));
-        return pokemonH2Repository.findAllBySpeciesId(speciesId)
-                .stream().map(factory.pokemontoPokemonResponse())
-                .collect(Collectors.toList());
+        if (securityContext.isCallerInRole(UserRole.ROLE_ADMIN)) {
+            return pokemonH2Repository.findAllBySpeciesId(speciesId)
+                    .stream().map(factory.pokemontoPokemonResponse())
+                    .collect(Collectors.toList());
+        }
+        else if (securityContext.isCallerInRole(UserRole.ROLE_USER)) {
+            return pokemonH2Repository.findAllBySpeciesIdAndUsername(speciesId, securityContext.getCallerPrincipal().getName())
+                    .stream().map(factory.pokemontoPokemonResponse())
+                    .collect(Collectors.toList());
+        }
+        else {
+            throw new EJBAccessException("Caller not authorized.");
+        }
     }
 
     public PokemonResponse findById(UUID id) {
+        checkAdminRoleOrOwner(pokemonH2Repository.find(id));
         return pokemonH2Repository.find(id)
                 .map(factory.pokemontoPokemonResponse())
                 .orElseThrow(() -> new NotFoundException("Pokemon with id " + id + " not found"));
@@ -98,11 +131,24 @@ public class PokemonService {
                 .orElseThrow(() -> new IllegalArgumentException("Pokemon with id " + id + " not found"));
     }
 
-    @Transactional
+
     public void delete(UUID id) {
+        checkAdminRoleOrOwner(pokemonH2Repository.find(id));
         Pokemon pokemon = pokemonH2Repository.find(id)
                 .orElseThrow(() -> new NotFoundException("Pokemon with id " + id + " not found"));
         pokemonH2Repository.delete(pokemon);
+    }
+
+    private void checkAdminRoleOrOwner(Optional<Pokemon> pokemon) throws EJBAccessException {
+        if (securityContext.isCallerInRole(UserRole.ROLE_ADMIN)) {
+            return;
+        }
+        if (securityContext.isCallerInRole(UserRole.ROLE_USER)
+                && pokemon.isPresent()
+                && pokemon.get().getOwner().getUsername().equals(securityContext.getCallerPrincipal().getName())) {
+            return;
+        }
+        throw new EJBAccessException("Caller not authorized.");
     }
 
 }
